Moved descriptor builder and filters to spark
- Use of a new SparkRouteData class
- Avoids having a dependency on Microsoft.AspNet.Mvc
