+++ src/Spark.Web.Mvc/Wrappers/ITextWriterContainer.cs
@@ -1,7 +1,9 @@
 ﻿using System.IO;
 
-namespace Spark.Web.Mvc.Wrappers {
-    public interface ITextWriterContainer {
-        TextWriter Output{ get; set;}
+namespace Spark.Web.Mvc.Wrappers
+{
+    public interface ITextWriterContainer
+    {
+        TextWriter Output { get; set; }
     }
 }
