+++ src/Spark/Compiler/BatchCompiler.cs
@@ -15,7 +15,6 @@ namespace Spark.Compiler;
 
 #if NETFRAMEWORK
 
-[Obsolete("To be replaced with RoslynBatchCompiler (code dom cannot target .net 'core')")]
 public class CodeDomBatchCompiler : IBatchCompiler
 {
     /// <summary>
